# api_clients/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"DefaultSelection_Prisma._36_ProfessorPayload_": {
				"properties": {
					"school_id": {
						"type": "integer",
						"format": "int32"
					},
					"rmp_id": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"first_name": {
						"type": "string"
					},
					"id": {
						"type": "integer",
						"format": "int32"
					}
				},
				"required": [
					"school_id",
					"rmp_id",
					"last_name",
					"first_name",
					"id"
				],
				"type": "object"
			},
			"Professor": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_ProfessorPayload_",
				"description": "Model Professor"
			},
...
			"DefaultSelection_Prisma._36_CoursePayload_": {
				"properties": {
					"school_id": {
						"type": "integer",
						"format": "int32"
					},
					"code": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "integer",
						"format": "int32"
					}
				},
				"required": [
					"school_id",
					"code",
					"name",
					"id"
				],
				"type": "object"
			},
			"Course": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_CoursePayload_",
				"description": "Model Course"
			},
			"_36_Enums.Semester": {
				"type": "string",
				"enum": [
					"SPRING",
					"SUMMER",
					"FALL"
				]
			},
			"Semester": {
				"$ref": "#/components/schemas/_36_Enums.Semester"
			},
...
			"CourseSearchResult": {
				"properties": {
					"pageInfo": {
						"properties": {
							"total": {
								"type": "integer",
								"format": "int32"
							},
							"endCursor": {
								"type": "string",
								"nullable": true
							},
							"hasNextPage": {
								"type": "boolean"
							}
						},
						"required": [
							"total",
							"endCursor",
							"hasNextPage"
						],
						"type": "object"
					},
					"edges": {
						"items": {
							"properties": {
								"node": {
									"$ref": "#/components/schemas/Course"
								},
								"cursor": {
									"type": "string"
								}
							},
							"required": [
								"node",
								"cursor"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"pageInfo",
					"edges"
				],
				"type": "object"
			},
			"CourseErrorType": {
				"enum": [
					"COURSE_NOT_FOUND",
					"COURSE_ALREADY_EXISTS"
				],
				"type": "string"
			},
			"Pick_CourseError.message-or-type_": {
				"properties": {
					"message": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"properties": {},
								"additionalProperties": {
									"type": "string"
								},
								"type": "object"
							}
						]
					},
					"type": {
						"$ref": "#/components/schemas/CourseErrorType"
					}
				},
				"required": [
					"message",
					"type"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"CourseErrorType.COURSE_ALREADY_EXISTS": {
				"enum": [
					"COURSE_ALREADY_EXISTS"
				],
				"type": "string"
			},
			"CourseAlreadyExistsError": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_CourseError.message-or-type_"
					},
					{
						"properties": {
							"type": {
								"$ref": "#/components/schemas/CourseErrorType.COURSE_ALREADY_EXISTS"
							}
						},
						"required": [
							"type"
						],
						"type": "object"
					}
				]
			},
			"Pick_Course.name-or-code-or-school_id_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"school_id": {
						"type": "integer",
						"format": "int32"
					},
					"code": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"school_id",
					"code"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"NewCourse": {
				"$ref": "#/components/schemas/Pick_Course.name-or-code-or-school_id_"
			},
			"CourseErrorType.COURSE_NOT_FOUND": {
				"enum": [
					"COURSE_NOT_FOUND"
				],
				"type": "string"
			},
			"CourseNotFoundError": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_CourseError.message-or-type_"
					},
					{
						"properties": {
							"type": {
								"$ref": "#/components/schemas/CourseErrorType.COURSE_NOT_FOUND"
							}
						},
						"required": [
							"type"
						],
						"type": "object"
					}
				]
			},
			"Pick_Course.name-or-code_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"code"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UpdatedCourse": {
				"$ref": "#/components/schemas/Pick_Course.name-or-code_"
			},
			"CourseProfessors": {
				"properties": {
					"total": {
						"type": "integer",
						"format": "int32"
					},
					"professors": {
						"items": {
							"$ref": "#/components/schemas/Professor"
						},
						"type": "array"
					}
				},
				"required": [
					"total",
					"professors"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/courses/search": {
			"get": {
				"operationId": "SearchCourses",
				"responses": {
					"200": {
						"description": "Courses Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CourseSearchResult"
								}
							}
						}
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "school_id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "semester",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Semester"
						}
					},
					{
						"in": "query",
						"name": "year",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "int32",
							"type": "integer"
						}
					}
				]
			}
		},
		"/courses/create": {
			"post": {
				"operationId": "CreateCourse",
				"responses": {
					"201": {
						"description": "Course Created Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Course"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 1,
											"name": "Example Course",
											"code": "EX101",
											"school_id": 1
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "Course already exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CourseAlreadyExistsError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewCourse"
							}
						}
					}
				}
			}
		},
		"/courses/{id}": {
			"get": {
				"operationId": "GetCourse",
				"responses": {
					"200": {
						"description": "Course Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Course"
								}
							}
						}
					},
					"404": {
						"description": "Course not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CourseNotFoundError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCourse",
				"responses": {
					"200": {
						"description": "Course Updated Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Course"
								}
							}
						}
					},
					"404": {
						"description": "Course not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CourseNotFoundError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatedCourse"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCourse",
				"responses": {
					"200": {
						"description": "Course Deleted Successfully"
					},
					"404": {
						"description": "Course not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CourseNotFoundError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				]
			}
		},
		"/courses/{id}/professors": {
			"get": {
				"operationId": "GetCourseProfessors",
				"responses": {
					"200": {
						"description": "Professors Retrieved Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CourseProfessors"
								}
							}
						}
					},
					"404": {
						"description": "Course not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CourseNotFoundError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Courses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "year",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "semester",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Semester"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}